parser/tests/tree_construction/basic.dat

#data
<a href=a>aa<marquee>aa<a href=b>bb</marquee>aa
#errors
(1,10): expected-doctype-but-got-start-tag
(1,45): end-tag-too-early
(1,47): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <a>
|       href="a"
|       "aa"
|       <marquee>
|         "aa"
|         <a>
|           href="b"
|           "bb"
|       "aa"
